v.0.0.1.2

+++ app/src/main/java/com/shurik/historymoment/db/Object.kt
@@ -1,7 +1,7 @@
 package com.shurik.historymoment.db
 
 //data class Coordinates(val latitude: Double? = 0.0, val longitude: Double? = 0.0)
-data class MapPosition(val type: String? = "", val coordinates: List<Int>? = null)
+data class MapPosition(val type: String? = "", val coordinates: List<Double>? = null)
 data class Address(val fulladdress: String? = "", val mapPosition: MapPosition? = null)
 data class Photo(val title: String? = "", val url: String? = "")
